Move AWS processor to own package
To simplify the code base the AWS EventBridge processor code is moved to its own (sub) package under the `processor` tree.

Changes were made to the methods to comply with the updated `processor` interface and changes to `map[string]*metrics.InvocationDetails`.
